refactor: auth config 변경
